remove reduction from label and move pipe and tank insulation to steam folder
