+++ src/app/calculator/steam/pipe-insulation-reduction/pipe-insulation-reduction-form/pipe-insulation-reduction-form.component.spec.ts
