+++ src/app/calculator/steam/pipe-insulation-reduction/pipe-insulation-reduction.component.spec.ts
