+++ src/app/calculator/steam/steam-list/steam-list.component.html
@@ -178,5 +178,33 @@
                 </p>
             </div>
         </div>
+        <div class="d-flex calc-item-row w-100">
+            <div class="p-2 col-2 d-flex justify-content-center align-items-center">
+                <div class="calc-icon-div d-flex justify-content-center align-items-center">
+                    <img src="assets/images/calculator-icons/utilities-icons/pipe-insulation-reduction-icon.png"
+                        class="calc-img" [routerLink]="['/calculators/pipe-insulation-reduction']">
+                </div>
+            </div>
+            <div class="d-flex flex-fill flex-column justify-content-center">
+                <a [routerLink]="['/calculators/pipe-insulation-reduction']">Pipe Insulation</a>
+                <p>
+                    This calculator is used to quantify the energy savings associated with insulating hot pipes.
+                </p>
+            </div>
+        </div>
+        <div class="d-flex calc-item-row w-100">
+            <div class="p-2 col-2 d-flex justify-content-center align-items-center">
+                <div class="calc-icon-div d-flex justify-content-center align-items-center">
+                    <img src="assets/images/calculator-icons/utilities-icons/tank-insulation-reduction-icon.png"
+                        class="calc-img" [routerLink]="['/calculators/tank-insulation-reduction']">
+                </div>
+            </div>
+            <div class="d-flex flex-fill flex-column justify-content-center">
+                <a [routerLink]="['/calculators/tank-insulation-reduction']">Tank Insulation</a>
+                <p>
+                    This calculator is used to quantify the energy savings associated with insulating hot tanks.
+                </p>
+            </div>
+        </div>
     </div>
 </div>
