+++ src/app/calculator/steam/steam.module.ts
@@ -16,6 +16,8 @@ import { BoilerBlowdownRateModule } from './boiler-blowdown-rate/boiler-blowdown
 import { SteamListComponent } from './steam-list/steam-list.component';
 import { RouterModule } from '@angular/router';
 import { SteamReductionModule } from './steam-reduction/steam-reduction.module';
+import { PipeInsulationReductionModule } from './pipe-insulation-reduction/pipe-insulation-reduction.module';
+import { TankInsulationReductionModule } from './tank-insulation-reduction/tank-insulation-reduction.module';
 
 
 @NgModule({
@@ -33,7 +35,9 @@ import { SteamReductionModule } from './steam-reduction/steam-reduction.module';
     TurbineModule,
     BoilerBlowdownRateModule,
     SteamReductionModule,
-    RouterModule
+    RouterModule,
+    PipeInsulationReductionModule,
+    TankInsulationReductionModule
   ],
   declarations: [
     SteamListComponent
