+++ src/app/calculator/steam/tank-insulation-reduction/tank-insulation-reduction-results/tank-insulation-reduction-results.component.css
