+++ src/app/calculator/steam/tank-insulation-reduction/tank-insulation-reduction.component.css
