+++ src/app/calculator/utilities/utilities-list/utilities-list.component.html
@@ -132,34 +132,6 @@
         <div class="d-flex calc-item-row w-100">
             <div class="p-2 col-2 d-flex justify-content-center align-items-center">
                 <div class="calc-icon-div d-flex justify-content-center align-items-center">
-                    <img src="assets/images/calculator-icons/utilities-icons/pipe-insulation-reduction-icon.png"
-                        class="calc-img" [routerLink]="['/calculators/pipe-insulation-reduction']">
-                </div>
-            </div>
-            <div class="d-flex flex-fill flex-column justify-content-center">
-                <a [routerLink]="['/calculators/pipe-insulation-reduction']">Pipe Insulation Reduction</a>
-                <p>
-                    This calculator is used to quantify the energy savings associated with insulating hot pipes.
-                </p>
-            </div>
-        </div>
-        <div class="d-flex calc-item-row w-100">
-            <div class="p-2 col-2 d-flex justify-content-center align-items-center">
-                <div class="calc-icon-div d-flex justify-content-center align-items-center">
-                    <img src="assets/images/calculator-icons/utilities-icons/tank-insulation-reduction-icon.png"
-                        class="calc-img" [routerLink]="['/calculators/tank-insulation-reduction']">
-                </div>
-            </div>
-            <div class="d-flex flex-fill flex-column justify-content-center">
-                <a [routerLink]="['/calculators/tank-insulation-reduction']">Tank Insulation Reduction</a>
-                <p>
-                    This calculator is used to quantify the energy savings associated with insulating hot tanks.
-                </p>
-            </div>
-        </div>
-        <div class="d-flex calc-item-row w-100">
-            <div class="p-2 col-2 d-flex justify-content-center align-items-center">
-                <div class="calc-icon-div d-flex justify-content-center align-items-center">
                     <img src="assets/images/calculator-icons/fan-icons/psychrometric-icon.png" class="calc-img"
                         [routerLink]="['/calculators/fan-psychrometric']">
                 </div>
