+++ src/app/calculator/utilities/utilities.module.ts
@@ -10,8 +10,6 @@ import { NaturalGasReductionModule } from './natural-gas-reduction/natural-gas-r
 import { WaterReductionModule } from './water-reduction/water-reduction.module';
 import { CashFlowModule } from './cash-flow/cash-flow.module';
 import { UnitConverterModule } from './unit-converter/unit-converter.module';
-import { PipeInsulationReductionModule } from './pipe-insulation-reduction/pipe-insulation-reduction.module';
-import { TankInsulationReductionModule } from './tank-insulation-reduction/tank-insulation-reduction.module';
 import { UtilitiesListComponent } from './utilities-list/utilities-list.component';
 import { RouterModule } from '@angular/router';
 import { FanPsychrometricModule } from './fan-psychrometric/fan-psychrometric.module';
@@ -34,8 +32,6 @@ import { FanPsychrometricModule } from './fan-psychrometric/fan-psychrometric.mo
         ElectricityReductionModule,
         NaturalGasReductionModule,
         WaterReductionModule,
-        PipeInsulationReductionModule,
-        TankInsulationReductionModule,
         CashFlowModule,
         UnitConverterModule,
         RouterModule,
