+++ src/app/dashboard/sidebar/calculator-list/calculator-list.component.html
@@ -1,4 +1,5 @@
-<span [routerLink]="['/calculators/calculators-list']" routerLinkActive="selected" class="sidebar-item">All Calculators</span>
+<span [routerLink]="['/calculators/calculators-list']" routerLinkActive="selected" class="sidebar-item">All
+  Calculators</span>
 <!--MOTORS-->
 <div routerLinkActive="show-border" [routerLinkActiveOptions]="{exact:true}">
   <span class="sidebar-item" [routerLink]="['/calculators/motors-list']" routerLinkActive="selected">Motors</span>
@@ -18,9 +19,11 @@
 </div>
 <!--process cooling-->
 <div routerLinkActive="show-border" [routerLinkActiveOptions]="{exact:true}">
-  <span class="sidebar-item" [routerLink]="['/calculators/process-cooling-list']" routerLinkActive="selected">Process Cooling</span>
+  <span class="sidebar-item" [routerLink]="['/calculators/process-cooling-list']" routerLinkActive="selected">Process
+    Cooling</span>
   <div class="directory-item">
-    <span [routerLink]="['/calculators/cooling-tower']" routerLinkActive="selected" class="sidebar-item">Cooling Tower Makeup Water</span>
+    <span [routerLink]="['/calculators/cooling-tower']" routerLinkActive="selected" class="sidebar-item">Cooling Tower
+      Makeup Water</span>
   </div>
 </div>
 <!--pumps-->
@@ -83,6 +86,10 @@
     <span [routerLink]="['/calculators/turbine']" routerLinkActive="selected" class="sidebar-item">Steam Turbine</span>
     <span [routerLink]="['/calculators/boiler-blowdown-rate']" routerLinkActive="selected" class="sidebar-item">Boiler
       Blowdown Rate</span>
+    <span [routerLink]="['/calculators/pipe-insulation-reduction']" routerLinkActive="selected"
+      class="sidebar-item">Pipe Insulation</span>
+    <span [routerLink]="['/calculators/tank-insulation-reduction']" routerLinkActive="selected"
+      class="sidebar-item">Tank Insulation</span>
   </div>
 </div>
 <!--compressed air-->
@@ -96,8 +103,8 @@
     <span [routerLink]="['/calculators/compressed-air-pressure-reduction']" routerLinkActive="selected"
       class="sidebar-item">Compressed Air
       Pressure Reduction</span>
-      <span [routerLink]="['/calculators/air-flow-conversion']" routerLinkActive="selected"
-      class="sidebar-item">Actual to Standard Airflow</span>
+    <span [routerLink]="['/calculators/air-flow-conversion']" routerLinkActive="selected" class="sidebar-item">Actual to
+      Standard Airflow</span>
     <span [routerLink]="['/calculators/bag-method']" routerLinkActive="selected" class="sidebar-item">Leak Loss
       Estimator - Bag Method</span>
     <span [routerLink]="['/calculators/pneumatic-air']" routerLinkActive="selected" class="sidebar-item">Pneumatic Air
@@ -114,7 +121,8 @@
       Capacity</span>
     <span [routerLink]="['/calculators/operating-cost']" routerLinkActive="selected" class="sidebar-item">Operating
       Costs</span>
-      <span [routerLink]="['/calculators/air-leak']" routerLinkActive="selected" class="sidebar-item">Air Leak Survey</span>
+    <span [routerLink]="['/calculators/air-leak']" routerLinkActive="selected" class="sidebar-item">Air Leak
+      Survey</span>
   </div>
 </div>
 <!--lighting-->
@@ -146,12 +154,8 @@
       Gas Reduction</span>
     <span [routerLink]="['/calculators/water-reduction']" routerLinkActive="selected" class="sidebar-item">Water
       Reduction</span>
-    <span [routerLink]="['/calculators/pipe-insulation-reduction']" routerLinkActive="selected"
-      class="sidebar-item">Pipe Insulation
-      Reduction</span>
-    <span [routerLink]="['/calculators/tank-insulation-reduction']" routerLinkActive="selected"
-      class="sidebar-item">Tank Insulation
-      Reduction</span>
-    <span [routerLink]="['/calculators/fan-psychrometric']" routerLinkActive="selected" class="sidebar-item">Psychrometric Calculator</span>
+
+    <span [routerLink]="['/calculators/fan-psychrometric']" routerLinkActive="selected"
+      class="sidebar-item">Psychrometric Calculator</span>
   </div>
 </div>
