+++ src/app/treasure-hunt/calculators/calculators.module.ts
@@ -19,7 +19,7 @@ import { GeneralDetailsFormComponent } from './opportunity-sheet/general-details
 import { StandaloneOpportunitySheetComponent } from './standalone-opportunity-sheet/standalone-opportunity-sheet.component';
 import { EnergyUseFormComponent } from './standalone-opportunity-sheet/energy-use-form/energy-use-form.component';
 import { FormsModule } from '@angular/forms';
-import { PipeInsulationReductionModule } from '../../calculator/utilities/pipe-insulation-reduction/pipe-insulation-reduction.module';
+import { PipeInsulationReductionModule } from '../../calculator/steam/pipe-insulation-reduction/pipe-insulation-reduction.module';
 import { CompressedAirReductionModule } from '../../calculator/compressed-air/compressed-air-reduction/compressed-air-reduction.module';
 import { CompressedAirPressureReductionModule } from '../../calculator/compressed-air/compressed-air-pressure-reduction/compressed-air-pressure-reduction.module';
 import { SteamReductionModule } from '../../calculator/steam/steam-reduction/steam-reduction.module';
