+++ src/app/treasure-hunt/find-treasure/find-treasure.component.html
@@ -185,7 +185,7 @@
         <div class="calc-icon d-flex justify-content-center align-items-center">
           <img src="assets/images/calculator-icons/utilities-icons/pipe-insulation-reduction-icon.png" class="calc-img">
         </div>
-        <a class="pt-2">Pipe Insulation Reduction</a>
+        <a class="pt-2">Pipe Insulation</a>
         <p class="mb-0 text-center help-text">
           This calculator is used to quantify the energy savings associated with insulating hot pipes.
         </p>
